tools/alerts/requirements.txt
pytest>=7.4
pyyaml>=6.0
prometheus-client>=0.20
